Remove timedelta directory, move more tests
