--- python/cudf/cudf/tests/indexes/timedelta/test_constructing.py
@@ -1,1 +1,0 @@
-# Copyright (c) 2023, NVIDIA CORPORATION.
