--- python/cudf/cudf/tests/indexes/timedelta/test_conversion.py
@@ -1,1 +1,0 @@
-# Copyright (c) 2023, NVIDIA CORPORATION.
